added a label htmlFor and some small adjustements
