+++ code/src/Components/Form.js
@@ -4,11 +4,11 @@ export const Form = ({ newMessage, onNewMessage, onFormSubmit }) => {
 
 return (
         <form className="posting-form" onSubmit={onFormSubmit}>
-            <h4>What's making you happy right now?</h4>
-            <textarea className="form-textarea"
-            value={newMessage} 
-            onChange={onNewMessage} 
-            placeholder='Type here...'
+            <label htmlfor="newMessage"><h4>What's making you happy right now?</h4></label>
+             <textarea className="form-textarea"
+                value={newMessage} 
+                onChange={onNewMessage} 
+                placeholder='Type here...'
             />
             <button className="form-button" type="submit" disabled={newMessage.length < 6 || newMessage.length > 140}>
             <span role="img" aria-label="heart">🖤 Send happy thought! 🖤</span>
